Implement AbstractGfx for Software & Null
