+++ Source/Core/VideoBackends/Software/SWGfx.h
@@ -1,19 +1,18 @@
-// Copyright 2015 Dolphin Emulator Project
+// Copyright 2023 Dolphin Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
 
-#include "VideoCommon/RenderBase.h"
+#include "VideoCommon/AbstractGfx.h"
 
-class BoundingBox;
+class SWOGLWindow;
 
-namespace Null
+namespace SW
 {
-class Renderer final : public ::Renderer
+class SWGfx final : public AbstractGfx
 {
 public:
-  Renderer();
-  ~Renderer() override;
+  SWGfx(std::unique_ptr<SWOGLWindow> window);
 
   bool IsHeadless() const override;
 
@@ -23,6 +22,8 @@ public:
   CreateStagingTexture(StagingTextureType type, const TextureConfig& config) override;
   std::unique_ptr<AbstractFramebuffer>
   CreateFramebuffer(AbstractTexture* color_attachment, AbstractTexture* depth_attachment) override;
+
+  void BindBackbuffer(const ClearColor& clear_color = {}) override;
 
   std::unique_ptr<AbstractShader> CreateShaderFromSource(ShaderStage stage, std::string_view source,
                                                          std::string_view name) override;
@@ -35,17 +36,20 @@ public:
                                                    const void* cache_data = nullptr,
                                                    size_t cache_data_length = 0) override;
 
-  u32 AccessEFB(EFBAccessType type, u32 x, u32 y, u32 poke_data) override { return 0; }
-  void PokeEFB(EFBAccessType type, const EfbPokeData* points, size_t num_points) override {}
+  void ShowImage(const AbstractTexture* source_texture,
+                 const MathUtil::Rectangle<int>& source_rc) override;
 
-  void ClearScreen(const MathUtil::Rectangle<int>& rc, bool colorEnable, bool alphaEnable,
-                   bool zEnable, u32 color, u32 z) override
-  {
-  }
+  void ScaleTexture(AbstractFramebuffer* dst_framebuffer, const MathUtil::Rectangle<int>& dst_rect,
+                    const AbstractTexture* src_texture,
+                    const MathUtil::Rectangle<int>& src_rect) override;
 
-  void ReinterpretPixelData(EFBReinterpretType convtype) override {}
+  void SetScissorRect(const MathUtil::Rectangle<int>& rc) override;
 
-protected:
-  std::unique_ptr<BoundingBox> CreateBoundingBox() const override;
+  void ClearRegion(const MathUtil::Rectangle<int>& rc, const MathUtil::Rectangle<int>& target_rc,
+                   bool colorEnable, bool alphaEnable, bool zEnable, u32 color, u32 z) override;
+
+private:
+  std::unique_ptr<SWOGLWindow> m_window;
 };
-}  // namespace Null
+
+}  // namespace SW
